requirements.txt
numpy~=1.23.3
pandas~=1.5.0
tensorflow~=2.10.0
opencv-python~=4.6.0.66
pytesseract~=0.3.10
Pillow~=9.2.0
keras~=2.10.0
matplotlib~=3.6.0
pdf2image~=1.16.0